+++ src/main/java/org/springframework/beans/context/ApplicationContext.java
@@ -1,0 +1,12 @@
+package org.springframework.beans.context;
+
+import org.springframework.beans.factory.ListableBeanFactory;
+
+/**
+ * 应用上下文接口
+ *
+ * @author: DoubleW2w
+ * @date: 2024/10/8
+ * @project: sbs-small-spring
+ */
+public interface ApplicationContext extends ListableBeanFactory {}