+++ src/main/java/org/springframework/beans/factory/config/BeanFactoryPostProcessor.java
@@ -1,0 +1,20 @@
+package org.springframework.beans.factory.config;
+
+import org.springframework.beans.BeansException;
+import org.springframework.beans.factory.ConfigurableListableBeanFactory;
+
+/**
+ * 允许在 Bean 对象注册后但未实例化之前，对 Bean 的定义信息 BeanDefinition 执行修改操作。
+ *
+ * @author derekyi
+ * @date 2020/11/28
+ */
+public interface BeanFactoryPostProcessor {
+  /**
+   * 在所有BeanDefinition加载完成后，但在bean实例化之前，提供修改BeanDefinition属性值的机制
+   *
+   * @param beanFactory bean工厂
+   * @throws BeansException 异常
+   */
+  void postProcessBeanFactory(ConfigurableListableBeanFactory beanFactory) throws BeansException;
+}