+++ src/main/java/org/springframework/beans/factory/support/AbstractAutowireCapableBeanFactory.java
@@ -6,7 +6,9 @@ import lombok.Setter;
 import org.springframework.beans.BeansException;
 import org.springframework.beans.PropertyValue;
 import org.springframework.beans.PropertyValues;
+import org.springframework.beans.factory.config.AutowireCapableBeanFactory;
 import org.springframework.beans.factory.config.BeanDefinition;
+import org.springframework.beans.factory.config.BeanPostProcessor;
 import org.springframework.beans.factory.config.BeanReference;
 
 import java.lang.reflect.Constructor;
@@ -18,19 +20,21 @@ import java.lang.reflect.Constructor;
  * @date: 2024/9/29
  * @project: sbs-small-spring
  */
-public abstract class AbstractAutowireCapableBeanFactory extends AbstractBeanFactory {
-
-  @Getter @Setter
+public abstract class AbstractAutowireCapableBeanFactory extends AbstractBeanFactory
+    implements AutowireCapableBeanFactory {
+  @Setter @Getter
   private InstantiationStrategy instantiationStrategy = new CglibSubclassingInstantiationStrategy();
 
   @Override
-  protected Object createBean(String beanName, BeanDefinition beanDefinition, Object[] args) {
-    Object bean;
+  protected Object createBean(String beanName, BeanDefinition beanDefinition, Object[] args)
+      throws BeansException {
+    Object bean = null;
     try {
-      // 创建Bean
       bean = createBeanInstance(beanDefinition, beanName, args);
-      // 填充属性
+      // 给 Bean 填充属性
       applyPropertyValues(beanName, bean, beanDefinition);
+      // 执行 Bean 的初始化方法和 BeanPostProcessor 的前置和后置处理方法
+      bean = initializeBean(beanName, bean, beanDefinition);
     } catch (Exception e) {
       throw new BeansException("Instantiation of bean failed", e);
     }
@@ -39,42 +43,44 @@ public abstract class AbstractAutowireCapableBeanFactory extends AbstractBeanFac
     return bean;
   }
 
-  /**
-   * 填充属性
-   *
-   * @param beanName bean名称
-   * @param bean bean对象
-   * @param beanDefinition bean定义
-   */
+  /** 初始化Bean */
+  private Object initializeBean(String beanName, Object bean, BeanDefinition beanDefinition) {
+    // 1. 执行 BeanPostProcessor Before 处理
+    Object wrappedBean = applyBeanPostProcessorsBeforeInitialization(bean, beanName);
+
+    // 待完成内容：invokeInitMethods(beanName, wrappedBean, beanDefinition);
+    invokeInitMethods(beanName, wrappedBean, beanDefinition);
+
+    // 2. 执行 BeanPostProcessor After 处理
+    wrappedBean = applyBeanPostProcessorsAfterInitialization(wrappedBean, beanName);
+    return wrappedBean;
+  }
+
+  private void invokeInitMethods(
+      String beanName, Object wrappedBean, BeanDefinition beanDefinition) {}
+
   private void applyPropertyValues(String beanName, Object bean, BeanDefinition beanDefinition) {
-    try{
+    try {
       PropertyValues propertyValues = beanDefinition.getPropertyValues();
       for (PropertyValue propertyValue : propertyValues.getPropertyValues()) {
+
         String name = propertyValue.getName();
         Object value = propertyValue.getValue();
-        // 如果是注入bean依赖
+
         if (value instanceof BeanReference) {
           // A 依赖 B，获取 B 的实例化
           BeanReference beanReference = (BeanReference) value;
           value = getBean(beanReference.getBeanName());
         }
-        // 为bean对象的name属性填充value
+        // 属性填充
         BeanUtil.setFieldValue(bean, name, value);
       }
-    }catch (Exception e){
+    } catch (Exception e) {
       throw new BeansException("Error setting property values：" + beanName);
-
     }
   }
 
-  /**
-   * 实例化Bean对象
-   *
-   * @param beanDefinition Bean定义
-   * @param beanName bean名称
-   * @param args 参数
-   * @return 结果
-   */
+  /** 创建Bean实例 */
   protected Object createBeanInstance(
       BeanDefinition beanDefinition, String beanName, Object[] args) {
     Constructor constructorToUse = null;
@@ -87,5 +93,29 @@ public abstract class AbstractAutowireCapableBeanFactory extends AbstractBeanFac
       }
     }
     return getInstantiationStrategy().instantiate(beanDefinition, beanName, constructorToUse, args);
+  }
+
+  @Override
+  public Object applyBeanPostProcessorsBeforeInitialization(Object existingBean, String beanName)
+      throws BeansException {
+    Object result = existingBean;
+    for (BeanPostProcessor processor : getBeanPostProcessors()) {
+      Object current = processor.postProcessBeforeInitialization(result, beanName);
+      if (null == current) return result;
+      result = current;
+    }
+    return result;
+  }
+
+  @Override
+  public Object applyBeanPostProcessorsAfterInitialization(Object existingBean, String beanName)
+      throws BeansException {
+    Object result = existingBean;
+    for (BeanPostProcessor processor : getBeanPostProcessors()) {
+      Object current = processor.postProcessAfterInitialization(result, beanName);
+      if (null == current) return result;
+      result = current;
+    }
+    return result;
   }
 }