+++ src/main/java/org/springframework/beans/factory/support/BeanDefinitionRegistry.java
@@ -11,6 +11,7 @@ import org.springframework.beans.factory.config.BeanDefinition;
  * @project: sbs-small-spring
  */
 public interface BeanDefinitionRegistry {
+
   /**
    * 向注册表中注册 BeanDefinition
    *
@@ -22,8 +23,8 @@ public interface BeanDefinitionRegistry {
   /**
    * 使用Bean名称查询BeanDefinition
    *
-   * @param beanName bean名称
-   * @return bean定义
+   * @param beanName
+   * @return
    * @throws BeansException
    */
   BeanDefinition getBeanDefinition(String beanName) throws BeansException;
@@ -31,8 +32,8 @@ public interface BeanDefinitionRegistry {
   /**
    * 判断是否包含指定名称的BeanDefinition
    *
-   * @param beanName bean名称
-   * @return 是否包含
+   * @param beanName
+   * @return
    */
   boolean containsBeanDefinition(String beanName);