+++ src/main/java/org/springframework/beans/factory/xml/XmlBeanDefinitionReader.java
@@ -33,57 +33,6 @@ public class XmlBeanDefinitionReader extends AbstractBeanDefinitionReader {
     super(registry, resourceLoader);
   }
 
-  protected void doLoadBeanDefinitions(InputStream inputStream) throws ClassNotFoundException {
-    Document doc = XmlUtil.readXML(inputStream);
-    Element root = doc.getDocumentElement();
-    NodeList childNodes = root.getChildNodes();
-    for (int i = 0; i < childNodes.getLength(); i++) {
-      // 判断元素
-      if (!(childNodes.item(i) instanceof Element)) continue;
-      // 判断对象
-      if (!"bean".equals(childNodes.item(i).getNodeName())) continue;
-
-      // 解析标签
-      Element bean = (Element) childNodes.item(i);
-      String id = bean.getAttribute("id");
-      String name = bean.getAttribute("name");
-      String className = bean.getAttribute("class");
-
-      // 获取 Class，方便获取类中的名称
-      Class<?> clazz = Class.forName(className);
-      // 优先级 id > name
-      String beanName = StrUtil.isNotBlank(id) ? id : name;
-      if (StrUtil.isNotBlank(beanName)) {
-        beanName = StrUtil.lowerFirst(clazz.getSimpleName());
-      }
-      // 定义Bean
-      BeanDefinition beanDefinition = new BeanDefinition(clazz);
-
-      // 读取属性并填充
-      NodeList beanChildNodes = bean.getChildNodes();
-      for (int j = 0; j < beanChildNodes.getLength(); j++) {
-        if (!(beanChildNodes.item(j) instanceof Element)) continue;
-        if (!"property".equals(beanChildNodes.item(j).getNodeName())) continue;
-        // 解析标签：property
-        Element property = (Element) beanChildNodes.item(j);
-        String attrName = property.getAttribute("name");
-        String attrValue = property.getAttribute("value");
-        String attrRef = property.getAttribute("ref");
-        // 获取属性值：引入对象、值对象
-        Object value = StrUtil.isNotBlank(attrRef) ? new BeanReference(attrRef) : attrValue;
-        // 创建属性信息
-        PropertyValue propertyValue = new PropertyValue(attrName, value);
-        beanDefinition.getPropertyValues().addPropertyValue(propertyValue);
-      }
-
-      if (getRegistry().containsBeanDefinition(beanName)) {
-        throw new BeansException("Duplicate beanName[" + beanName + "] is not allowed");
-      }
-      // 注册 BeanDefinition
-      getRegistry().registerBeanDefinition(beanName, beanDefinition);
-    }
-  }
-
   @Override
   public void loadBeanDefinitions(Resource resource) throws BeansException {
     try {
@@ -107,5 +56,61 @@ public class XmlBeanDefinitionReader extends AbstractBeanDefinitionReader {
     ResourceLoader resourceLoader = getResourceLoader();
     Resource resource = resourceLoader.getResource(location);
     loadBeanDefinitions(resource);
+  }
+
+  @Override
+  public void loadBeanDefinitions(String... locations) throws BeansException {
+    for (String location : locations) {
+      loadBeanDefinitions(location);
+    }
+  }
+
+  protected void doLoadBeanDefinitions(InputStream inputStream) throws ClassNotFoundException {
+    Document doc = XmlUtil.readXML(inputStream);
+    Element root = doc.getDocumentElement();
+    NodeList childNodes = root.getChildNodes();
+
+    for (int i = 0; i < childNodes.getLength(); i++) {
+      // 判断元素
+      if (!(childNodes.item(i) instanceof Element)) continue;
+      // 判断对象
+      if (!"bean".equals(childNodes.item(i).getNodeName())) continue;
+
+      // 解析标签
+      Element bean = (Element) childNodes.item(i);
+      String id = bean.getAttribute("id");
+      String name = bean.getAttribute("name");
+      String className = bean.getAttribute("class");
+      // 获取 Class，方便获取类中的名称
+      Class<?> clazz = Class.forName(className);
+      // 优先级 id > name
+      String beanName = StrUtil.isNotEmpty(id) ? id : name;
+      if (StrUtil.isEmpty(beanName)) {
+        beanName = StrUtil.lowerFirst(clazz.getSimpleName());
+      }
+
+      // 定义Bean
+      BeanDefinition beanDefinition = new BeanDefinition(clazz);
+      // 读取属性并填充
+      for (int j = 0; j < bean.getChildNodes().getLength(); j++) {
+        if (!(bean.getChildNodes().item(j) instanceof Element)) continue;
+        if (!"property".equals(bean.getChildNodes().item(j).getNodeName())) continue;
+        // 解析标签：property
+        Element property = (Element) bean.getChildNodes().item(j);
+        String attrName = property.getAttribute("name");
+        String attrValue = property.getAttribute("value");
+        String attrRef = property.getAttribute("ref");
+        // 获取属性值：引入对象、值对象
+        Object value = StrUtil.isNotEmpty(attrRef) ? new BeanReference(attrRef) : attrValue;
+        // 创建属性信息
+        PropertyValue propertyValue = new PropertyValue(attrName, value);
+        beanDefinition.getPropertyValues().addPropertyValue(propertyValue);
+      }
+      if (getRegistry().containsBeanDefinition(beanName)) {
+        throw new BeansException("Duplicate beanName[" + beanName + "] is not allowed");
+      }
+      // 注册 BeanDefinition
+      getRegistry().registerBeanDefinition(beanName, beanDefinition);
+    }
   }
 }