+++ src/test/java/org/springframework/beans/factory/ApplicationContextPostProcessorTest.java
@@ -1,0 +1,48 @@
+package org.springframework.beans.factory;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.context.support.ClassPathXmlApplicationContext;
+import org.springframework.beans.factory.support.DefaultListableBeanFactory;
+import org.springframework.beans.factory.xml.XmlBeanDefinitionReader;
+
+/**
+ * @author: DoubleW2w
+ * @date: 2024/10/8
+ * @project: sbs-small-spring
+ */
+public class ApplicationContextPostProcessorTest {
+  @Test
+  public void test_BeanFactoryPostProcessorAndBeanPostProcessor() {
+    // 1.初始化 BeanFactory
+    DefaultListableBeanFactory beanFactory = new DefaultListableBeanFactory();
+
+    // 2. 读取配置文件&注册Bean
+    XmlBeanDefinitionReader reader = new XmlBeanDefinitionReader(beanFactory);
+    reader.loadBeanDefinitions("classpath:spring-config.xml");
+
+    // 3. BeanDefinition 加载完成 & Bean实例化之前，修改 BeanDefinition 的属性值
+    MyBeanFactoryPostProcessor beanFactoryPostProcessor = new MyBeanFactoryPostProcessor();
+    beanFactoryPostProcessor.postProcessBeanFactory(beanFactory);
+
+    // 4. Bean实例化之后，修改 Bean 属性信息
+    MyBeanPostProcessor beanPostProcessor = new MyBeanPostProcessor();
+    beanFactory.addBeanPostProcessor(beanPostProcessor);
+
+    // 5. 获取Bean对象调用方法
+    HelloService helloService = beanFactory.getBean("helloService", HelloService.class);
+    String result = helloService.sayHello();
+    System.out.println("测试结果：" + result);
+  }
+
+  @Test
+  public void test_ApplicationContext() {
+    // 1.初始化 BeanFactory
+    ClassPathXmlApplicationContext applicationContext =
+        new ClassPathXmlApplicationContext("classpath:spring-config.xml");
+
+    // 2. 获取Bean对象调用方法
+    HelloService helloService = applicationContext.getBean("helloService", HelloService.class);
+    String result = helloService.sayHello();
+    System.out.println("测试结果：" + result);
+  }
+}