+++ src/test/java/org/springframework/beans/factory/MyBeanFactoryPostProcessor.java
@@ -1,0 +1,23 @@
+package org.springframework.beans.factory;
+
+import org.springframework.beans.BeansException;
+import org.springframework.beans.PropertyValue;
+import org.springframework.beans.PropertyValues;
+import org.springframework.beans.factory.config.BeanDefinition;
+import org.springframework.beans.factory.config.BeanFactoryPostProcessor;
+
+/**
+ * @author: DoubleW2w
+ * @date: 2024/10/8
+ * @project: sbs-small-spring
+ */
+public class MyBeanFactoryPostProcessor implements BeanFactoryPostProcessor {
+  @Override
+  public void postProcessBeanFactory(ConfigurableListableBeanFactory beanFactory)
+      throws BeansException {
+    BeanDefinition beanDefinition = beanFactory.getBeanDefinition("helloService");
+    PropertyValues propertyValues = beanDefinition.getPropertyValues();
+
+    propertyValues.addPropertyValue(new PropertyValue("name", "改为：字节跳动"));
+  }
+}