+++ src/test/java/org/springframework/beans/factory/MyBeanPostProcessor.java
@@ -1,0 +1,27 @@
+package org.springframework.beans.factory;
+
+import org.springframework.beans.BeansException;
+import org.springframework.beans.factory.config.BeanPostProcessor;
+
+/**
+ * @author: DoubleW2w
+ * @date: 2024/10/8
+ * @project: sbs-small-spring
+ */
+public class MyBeanPostProcessor implements BeanPostProcessor {
+
+  @Override
+  public Object postProcessBeforeInitialization(Object bean, String beanName)
+      throws BeansException {
+    if ("helloService".equals(beanName)) {
+      HelloService helloService = (HelloService) bean;
+      helloService.setName("改为：北京");
+    }
+    return bean;
+  }
+
+  @Override
+  public Object postProcessAfterInitialization(Object bean, String beanName) throws BeansException {
+    return bean;
+  }
+}